+++ source/Relativity.Testing.Framework/Models/Matter/Matter.cs
@@ -1,9 +1,11 @@
-﻿namespace Relativity.Testing.Framework.Models
+﻿using System.Collections.Generic;
+
+namespace Relativity.Testing.Framework.Models
 {
 	/// <summary>
 	/// Represents the Relativity matter object.
 	/// </summary>
-	public class Matter : NamedArtifact, IFillsRequiredProperties<Matter>
+	public class Matter : TimeStampedNamedArtifact, IFillsRequiredProperties<Matter>
 	{
 		/// <summary>
 		/// Gets or sets the number.
@@ -29,6 +31,11 @@
 		/// Gets or sets the notes.
 		/// </summary>
 		public string Notes { get; set; } = string.Empty;
+
+		/// <summary>
+		/// Gets or sets a list of RESTful operations that a user has permissions to perform on the matter.
+		/// </summary>
+		public List<Action> Actions { get; set; }
 
 		/// <summary>
 		/// Fills the required properties.