+++ source/Relativity.Testing.Framework/Models/Scripts/ActionJob.cs
@@ -59,7 +59,7 @@ namespace Relativity.Testing.Framework.Models
 		/// <summary>
 		/// Gets or sets a list of RESTful operations that a user has permissions to perform on the script action job.
 		/// </summary>
-		public List<ScriptAction> Actions { get; set; }
+		public List<Action> Actions { get; set; }
 
 		/// <summary>
 		/// Gets or sets a list of unsupported and read-only properties on the script action job.